HEEDLS-862 Move Remove from Course actions to be with other similar methods
