+++ DigitalLearningSolutions.Web.Tests/Controllers/TrackingSystem/Delegates/DelegateProgressControllerTests.cs
@@ -7,12 +7,14 @@
     using DigitalLearningSolutions.Data.Models.CustomPrompts;
     using DigitalLearningSolutions.Data.Models.Progress;
     using DigitalLearningSolutions.Data.Models.SearchSortFilterPaginate;
+    using DigitalLearningSolutions.Data.Models.User;
     using DigitalLearningSolutions.Data.Services;
     using DigitalLearningSolutions.Data.Tests.TestHelpers;
     using DigitalLearningSolutions.Web.Controllers.TrackingSystem.Delegates;
     using DigitalLearningSolutions.Web.Models.Enums;
     using DigitalLearningSolutions.Web.Tests.ControllerHelpers;
     using DigitalLearningSolutions.Web.ViewModels.TrackingSystem.Delegates.DelegateProgress;
+    using DigitalLearningSolutions.Web.ViewModels.TrackingSystem.Delegates.ViewDelegate;
     using FakeItEasy;
     using FluentAssertions.AspNetCore.Mvc;
     using Microsoft.Extensions.Configuration;
@@ -337,6 +339,160 @@
                 : null;
             result.Should().BeRedirectToActionResult().WithControllerName(expectedController)
                 .WithActionName(expectedAction).WithFragment(expectedFragment);
+        }
+
+        [Test]
+        public void Removal_confirmation_page_displays_for_valid_delegate_and_course()
+        {
+            // Given
+            A.CallTo(() => userService.GetDelegateUserCardById(1))
+                .Returns(new DelegateUserCard { CentreId = 2, Id = 1 });
+
+            A.CallTo(() => courseService.GetCourseNameAndApplication(1))
+                .Returns(new CourseNameInfo());
+
+            A.CallTo(() => courseService.DelegateHasCurrentProgress(1, 1))
+                .Returns(true);
+
+            // When
+            var result = delegateProgressController.ConfirmRemoveFromCourse(
+                1,
+                1,
+                DelegateAccessRoute.ViewDelegate
+            );
+
+            // Then
+            result.Should().BeViewResult();
+        }
+
+        [Test]
+        public void Removal_confirmation_page_returns_not_found_result_for_delegate_with_no_active_progress()
+        {
+            // Given
+            A.CallTo(() => userService.GetDelegateUserCardById(1))
+                .Returns(new DelegateUserCard { CentreId = 2, Id = 1 });
+
+            A.CallTo(() => courseService.GetCourseNameAndApplication(1))
+                .Returns(new CourseNameInfo());
+
+            A.CallTo(() => courseService.DelegateHasCurrentProgress(1, 1))
+                .Returns(false);
+
+            // When
+            var result = delegateProgressController.ConfirmRemoveFromCourse(
+                1,
+                1,
+                DelegateAccessRoute.ViewDelegate
+            );
+
+            // Then
+            result.Should().BeNotFoundResult();
+        }
+
+        [Test]
+        public void Delegate_removal_for_delegate_with_no_active_progress_returns_not_found_result()
+        {
+            // Given
+            A.CallTo(() => userService.GetDelegateUserCardById(1))
+                .Returns(new DelegateUserCard { CentreId = 2, Id = 1 });
+
+            A.CallTo(() => courseService.GetCourseNameAndApplication(1))
+                .Returns(new CourseNameInfo());
+
+            A.CallTo(() => courseService.DelegateHasCurrentProgress(1, 1))
+                .Returns(false);
+
+            var model = new RemoveFromCourseViewModel
+            {
+                Confirm = true,
+                AccessedVia = DelegateAccessRoute.ViewDelegate,
+            };
+
+            // When
+            var result = delegateProgressController.ExecuteRemoveFromCourse(
+                1,
+                1,
+                DelegateAccessRoute.ViewDelegate,
+                model
+            );
+
+            // Then
+            result.Should().BeNotFoundResult();
+        }
+
+        [Test]
+        public void Delegate_removal_for_valid_delegate_returns_redirect_to_action_view_delegate()
+        {
+            // Given
+            var delegateId = 1;
+
+            A.CallTo(() => userService.GetDelegateUserCardById(1))
+                .Returns(new DelegateUserCard { CentreId = 2, Id = 1 });
+
+            A.CallTo(() => courseService.GetCourseNameAndApplication(1))
+                .Returns(new CourseNameInfo());
+
+            A.CallTo(() => courseService.DelegateHasCurrentProgress(1, 1))
+                .Returns(true);
+
+            var model = new RemoveFromCourseViewModel
+            {
+                Confirm = true,
+                AccessedVia = DelegateAccessRoute.ViewDelegate,
+            };
+
+            // When
+            var result = delegateProgressController.ExecuteRemoveFromCourse(
+                delegateId,
+                1,
+                DelegateAccessRoute.ViewDelegate,
+                model
+            );
+
+            // Then
+            result.Should()
+                .BeRedirectToActionResult()
+                .WithActionName("Index")
+                .WithControllerName("ViewDelegate")
+                .WithRouteValue("delegateId", delegateId);
+        }
+
+        [Test]
+        public void Delegate_removal_for_valid_delegate_returns_redirect_to_action_delegate_courses()
+        {
+            // Given
+            var customisationId = 1;
+
+            A.CallTo(() => userService.GetDelegateUserCardById(1))
+                .Returns(new DelegateUserCard { CentreId = 2, Id = 1 });
+
+            A.CallTo(() => courseService.GetCourseNameAndApplication(1))
+                .Returns(new CourseNameInfo());
+
+            A.CallTo(() => courseService.DelegateHasCurrentProgress(1, 1))
+                .Returns(true);
+
+            var model = new RemoveFromCourseViewModel
+            {
+                Confirm = true,
+                AccessedVia = DelegateAccessRoute.CourseDelegates,
+                ReturnPageQuery = ReturnPageQueryHelper.GetDefaultReturnPageQuery(),
+            };
+
+            // When
+            var result = delegateProgressController.ExecuteRemoveFromCourse(
+                1,
+                customisationId,
+                DelegateAccessRoute.CourseDelegates,
+                model
+            );
+
+            // Then
+            result.Should()
+                .BeRedirectToActionResult()
+                .WithActionName("Index")
+                .WithControllerName("CourseDelegates")
+                .WithRouteValue("customisationId", customisationId.ToString());
         }
     }
 }
