+++ DigitalLearningSolutions.Web/Views/TrackingSystem/Delegates/Shared/_DelegateCourseProgressButtons.cshtml
@@ -11,16 +11,16 @@
 </a>
 
 @if (string.IsNullOrEmpty(Model.Completed) && Model.RemovedDate == null) {
-  <a
-    class="nhsuk-button delete-button"
-    role="button"
-    data-return-page-enabled="true"
-    asp-controller="ViewDelegate"
-    asp-action="ConfirmRemoveFromCourse"
-    asp-route-delegateId="@Model.DelegateId"
-    asp-route-customisationId="@Model.CustomisationId"
-    asp-route-accessedVia="@Model.AccessedVia"
-    asp-route-returnPageQuery="@Model.ReturnPageQuery">
+  <a class="nhsuk-button delete-button"
+     role="button"
+     data-return-page-enabled="true"
+     asp-controller="DelegateProgress"
+     asp-action="ConfirmRemoveFromCourse"
+     asp-route-progressId="@Model.ProgressId"
+     asp-route-delegateId="@Model.DelegateId"
+     asp-route-customisationId="@Model.CustomisationId"
+     asp-route-accessedVia="@Model.AccessedVia"
+     asp-route-returnPageQuery="@Model.ReturnPageQuery">
     Remove from course
   </a>
 }
